Allow BigInteger ethereum chain ids, improve handling of chain switching edge cases

+++ src/Reown.Sign.Nethereum/Runtime/Model/EthereumChain.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using Newtonsoft.Json;
 using Reown.Core.Common.Utils;
 
@@ -33,7 +34,7 @@ namespace Reown.Sign.Nethereum.Model
         public EthereumChain(string chainIdDecimal, string name, in Currency nativeCurrency, string[] rpcUrls, string[] blockExplorerUrls = null)
         {
             this.chainIdDecimal = chainIdDecimal;
-            chainIdHex = int.Parse(chainIdDecimal).ToHex(true);
+            chainIdHex = BigInteger.Parse(chainIdDecimal).ToHex(true);
             this.name = name;
             this.nativeCurrency = nativeCurrency;
             this.rpcUrls = rpcUrls;
